emergency update to limit coins prevent DoS
Emergency update apply patch to limit chain to a total of 7.5 billion blocks. No fork required, just upgrade. If patch not applied in 14 years there will be >7.5b coins. this will be discussed on btctalk thread shortly.

this issue was first noted on dogecoin but unfortunately was missed on the first round of coding. This causes no fatal errors and the network is young.